dr_api/requirements.txt
tensorflow==2.13.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pillow==10.1.0
numpy==1.24.3
pydantic==2.5.0